// packages/buildkite/src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export enum BuildStates {
  /** The build is blocked */
  Blocked = 'BLOCKED',
  /** The build was canceled */
  Canceled = 'CANCELED',
  /** The build is currently being canceled */
  Canceling = 'CANCELING',
  /** The build is currently being created */
  Creating = 'CREATING',
  /** The build failed */
  Failed = 'FAILED',
  /** The build is failing */
  Failing = 'FAILING',
  /** The build wasn't run */
  NotRun = 'NOT_RUN',
  /** The build passed */
  Passed = 'PASSED',
  /** The build is currently running jobs */
  Running = 'RUNNING',
  /** The build has yet to start running jobs */
  Scheduled = 'SCHEDULED',
  /** The build was skipped */
  Skipped = 'SKIPPED'
}
// ...
export type JobInterface = {
  /** If this job has been retried */
  readonly retried: Scalars['Boolean']['output'];
  /** The user that retried this job */
  readonly retriedBy?: Maybe<User>;
  /** The number of times the job has been retried */
  readonly retriesCount?: Maybe<Scalars['Int']['output']>;
  /** The job that was retried to create this job */
  readonly retrySource?: Maybe<Job>;
  /** The type of retry that was performed on this job */
  readonly retryType?: Maybe<JobRetryTypes>;
  /** The UUID for this job */
  readonly uuid: Scalars['String']['output'];
};
// ...
export enum JobStates {
  /** The job was accepted by the agent, and now it's waiting to start running */
  Accepted = 'ACCEPTED',
  /** The job has been assigned to an agent, and it's waiting for it to accept */
  Assigned = 'ASSIGNED',
  /** The job is waiting on a `block` step to finish */
  Blocked = 'BLOCKED',
  /** The job was in a `BLOCKED` state when the build failed */
  BlockedFailed = 'BLOCKED_FAILED',
  /** The jobs configuration means that it can't be run */
  Broken = 'BROKEN',
  /** The job was canceled */
  Canceled = 'CANCELED',
  /** The job is currently canceling */
  Canceling = 'CANCELING',
  /** The job expired before it was started on an agent */
  Expired = 'EXPIRED',
  /** The job has finished */
  Finished = 'FINISHED',
  /** The job is waiting for jobs with the same concurrency group to finish */
  Limited = 'LIMITED',
  /** The job is waiting on a concurrency group check before becoming either `LIMITED` or `SCHEDULED` */
  Limiting = 'LIMITING',
  /** The job has just been created and doesn't have a state yet */
  Pending = 'PENDING',
  /** The job is blocked by platform-wide concurrency/resource limits */
  PlatformLimited = 'PLATFORM_LIMITED',
  /** The job is waiting on platform-wide concurrency/resource limits before becoming either `PLATFORM_LIMITED` or `SCHEDULED` */
  PlatformLimiting = 'PLATFORM_LIMITING',
  /** The job is reserved by a stack for later execution */
  Reserved = 'RESERVED',
  /** The job is running */
  Running = 'RUNNING',
  /** The job is scheduled and waiting for an agent */
  Scheduled = 'SCHEDULED',
  /** The job was skipped */
  Skipped = 'SKIPPED',
  /** The job timed out */
  TimedOut = 'TIMED_OUT',
  /** The job is timing out for taking too long */
  TimingOut = 'TIMING_OUT',
  /** This `block` job has been manually unblocked */
  Unblocked = 'UNBLOCKED',
  /** This `block` job was in an `UNBLOCKED` state when the build failed */
  UnblockedFailed = 'UNBLOCKED_FAILED',
  /** The job is waiting on a `wait` step to finish */
  Waiting = 'WAITING',
  /** The job was in a `WAITING` state when the build failed */
  WaitingFailed = 'WAITING_FAILED'
}

/** Searching for jobs based on step information */
export type JobStepSearch = {
  /** The key assigned to the step */
  readonly key?: InputMaybe<ReadonlyArray<Scalars['String']['input']>>;
};

/** A type of job that requires a user to unblock it before proceeding in a build pipeline */
export type JobTypeBlock = JobInterface & Node & {
  /** The build that this job is a part of */
  readonly build?: Maybe<Build>;
  readonly id: Scalars['ID']['output'];
  /** Whether or not this job can be unblocked yet (may be waiting on another job to finish) */
  readonly isUnblockable?: Maybe<Scalars['Boolean']['output']>;
  /** The label of this block step */
  readonly label?: Maybe<Scalars['String']['output']>;
  /** If this job has been retried */
  readonly retried: Scalars['Boolean']['output'];
  /** The user that retried this job */
  readonly retriedBy?: Maybe<User>;
  /** The number of times the job has been retried */
  readonly retriesCount?: Maybe<Scalars['Int']['output']>;
  /** The job that was retried to create this job */
  readonly retrySource?: Maybe<Job>;
  /** The type of retry that was performed on this job */
  readonly retryType?: Maybe<JobRetryTypes>;
  /** The state of the job */
  readonly state: JobStates;
  /** The step that defined this job. Some older jobs in the system may not have an associated step */
  readonly step?: Maybe<StepInput>;
  /** The time when the job was created */
  readonly unblockedAt?: Maybe<Scalars['DateTime']['output']>;
  /** The user that unblocked this job */
  readonly unblockedBy?: Maybe<User>;
  /** The UUID for this job */
  readonly uuid: Scalars['String']['output'];
};
// ...
export type JobTypeCommand = JobInterface & Node & {
  /** The agent that is running the job */
  readonly agent?: Maybe<Agent>;
  /** The ruleset used to find an agent to run this job */
  readonly agentQueryRules?: Maybe<ReadonlyArray<Scalars['String']['output']>>;
  readonly annotations?: Maybe<AnnotationConnection>;
  /** Artifacts uploaded to this job */
  readonly artifacts?: Maybe<ArtifactConnection>;
  /** A glob of files to automatically upload after the job finishes */
  readonly automaticArtifactUploadPaths?: Maybe<Scalars['String']['output']>;
  /** The build that this job is a part of */
  readonly build?: Maybe<Build>;
  /** The time when the job was cancelled */
  readonly canceledAt?: Maybe<Scalars['DateTime']['output']>;
  /** The cluster of this job */
  readonly cluster?: Maybe<Cluster>;
  /** The cluster queue of this job */
  readonly clusterQueue?: Maybe<ClusterQueue>;
  /** The command the job will run */
  readonly command?: Maybe<Scalars['String']['output']>;
  /** Concurrency information related to a job */
  readonly concurrency?: Maybe<JobConcurrency>;
  /** The time when the job was created */
  readonly createdAt?: Maybe<Scalars['DateTime']['output']>;
  /** Environment variables for this job */
  readonly env?: Maybe<ReadonlyArray<Scalars['String']['output']>>;
  /** Job events */
  readonly events: JobEventConnection;
  /** The exit status returned by the command on the agent */
  readonly exitStatus?: Maybe<Scalars['String']['output']>;
  /** The time when the job was expired */
  readonly expiredAt?: Maybe<Scalars['DateTime']['output']>;
  /** The time when the job finished */
  readonly finishedAt?: Maybe<Scalars['DateTime']['output']>;
  readonly id: Scalars['ID']['output'];
  /** The label of the job */
  readonly label?: Maybe<Scalars['String']['output']>;
  /** The matrix configuration values for this particular job */
  readonly matrix?: Maybe<Scalars['JSON']['output']>;
  /** The index of this job within the parallel job group it is a part of. Null if this job is not part of a parallel job group. */
  readonly parallelGroupIndex?: Maybe<Scalars['Int']['output']>;
  /** The total number of jobs in the parallel job group this job is a part of. Null if this job is not part of a parallel job group. */
  readonly parallelGroupTotal?: Maybe<Scalars['Int']['output']>;
  /** If the job has finished and passed */
  readonly passed: Scalars['Boolean']['output'];
  /** The pipeline that this job is a part of */
  readonly pipeline?: Maybe<Pipeline>;
  /** The priority of this job */
  readonly priority: JobPriority;
  /** If this job has been retried */
  readonly retried: Scalars['Boolean']['output'];
  /** The user that retried this job */
  readonly retriedBy?: Maybe<User>;
  /** The number of times the job has been retried */
  readonly retriesCount?: Maybe<Scalars['Int']['output']>;
  /** Job retry rules */
  readonly retryRules?: Maybe<JobRetryRules>;
  /** The job that was retried to create this job */
  readonly retrySource?: Maybe<Job>;
  /** The type of retry that was performed on this job */
  readonly retryType?: Maybe<JobRetryTypes>;
  /** The time when the job became available to be run by an agent */
  readonly runnableAt?: Maybe<Scalars['DateTime']['output']>;
  /** The time when the job became scheduled for running */
  readonly scheduledAt?: Maybe<Scalars['DateTime']['output']>;
  /** The termination signal which killed the command, if the command was killed */
  readonly signal?: Maybe<Scalars['String']['output']>;
  /** If the termination signal was sent by the agent, the reason the agent took that action. If this field is null, and the `signal` field is not null, the command was killed by another process or by the operating system. */
  readonly signalReason?: Maybe<JobEventSignalReason>;
  /** If the job soft failed */
  readonly softFailed: Scalars['Boolean']['output'];
  /** The time when the job started running */
  readonly startedAt?: Maybe<Scalars['DateTime']['output']>;
  /** The state of the job */
  readonly state: JobStates;
  /** The step that defined this job. Some older jobs in the system may not have an associated step */
  readonly step?: Maybe<StepCommand>;
  /** The URL for the job */
  readonly url: Scalars['String']['output'];
  /** The UUID for this job */
  readonly uuid: Scalars['String']['output'];
};
// ...
export type JobTypeTrigger = JobInterface & Node & {
  /** Whether the triggered build runs asynchronously or not */
  readonly async: Scalars['Boolean']['output'];
  /** The build that this job is a part of */
  readonly build?: Maybe<Build>;
  readonly id: Scalars['ID']['output'];
  /** The label of this trigger step */
  readonly label?: Maybe<Scalars['String']['output']>;
  /** If this job has been retried */
  readonly retried: Scalars['Boolean']['output'];
  /** The user that retried this job */
  readonly retriedBy?: Maybe<User>;
  /** The number of times the job has been retried */
  readonly retriesCount?: Maybe<Scalars['Int']['output']>;
  /** The job that was retried to create this job */
  readonly retrySource?: Maybe<Job>;
  /** The type of retry that was performed on this job */
  readonly retryType?: Maybe<JobRetryTypes>;
  /** The state of the job */
  readonly state: JobStates;
  /** The step that defined this job. Some older jobs in the system may not have an associated step */
  readonly step?: Maybe<StepTrigger>;
  /** The build that this job triggered */
  readonly triggered?: Maybe<Build>;
  /** The UUID for this job */
  readonly uuid: Scalars['String']['output'];
};
// ...
export type Node = {
  /** ID of the object. */
  readonly id: Scalars['ID']['output'];
};
// ...
export type BuildDetailsQuery = { readonly build?: { readonly id: string, readonly number: number, readonly state: BuildStates, readonly url: string, readonly message?: string | null, readonly branch: string, readonly commit: string, readonly createdAt?: string | null, readonly startedAt?: string | null, readonly finishedAt?: string | null, readonly rebuiltFrom?: { readonly number: number } | null, readonly pipeline: { readonly name: string, readonly slug: string }, readonly jobs?: { readonly edges?: ReadonlyArray<{ readonly node?:
          | { readonly __typename: 'JobTypeBlock', readonly id: string, readonly uuid: string, readonly label?: string | null, readonly state: JobStates, readonly isUnblockable?: boolean | null }
          | { readonly __typename: 'JobTypeCommand', readonly id: string, readonly uuid: string, readonly label?: string | null, readonly state: JobStates, readonly url: string, readonly startedAt?: string | null, readonly finishedAt?: string | null, readonly softFailed: boolean, readonly exitStatus?: string | null, readonly retried: boolean, readonly retriesCount?: number | null, readonly parallelGroupIndex?: number | null, readonly parallelGroupTotal?: number | null }
          | { readonly __typename: 'JobTypeTrigger', readonly uuid: string, readonly label?: string | null, readonly state: JobStates }
          | { readonly __typename: 'JobTypeWait', readonly uuid: string, readonly state: JobStates }
         | null } | null> | null } | null } | null };